Ajeitando calculadora e rotas

+++ calcarb/forms.py
@@ -31,7 +31,7 @@ def veic_validator(valor):
 class Perguntas(forms.Form):
     precMes = forms.FloatField(widget=forms.NumberInput(attrs={'placeholder':'Responda em R$','class':'form-control'})) # x 0,0385 mensal
     precRegion = forms.FloatField(widget=forms.NumberInput(attrs={'placeholder':'Responda em R$','class':'form-control'})) # preço cobrado por kwh
-    gasCoz = forms.IntegerField(widget=forms.NumberInput(attrs={'placeholder':'Responda aqui','class':'form-control'})) # x 2,9380 mensal
+    gasCoz = forms.FloatField(widget=forms.NumberInput(attrs={'placeholder':'Responda aqui','class':'form-control'})) # x 2,9380 mensal
     veiculoP = forms.ChoiceField(validators=[veic_validator],choices=TRANSPORTE_VEIC,widget=forms.Select(attrs={'class':'seletor form-control'})) #
     tCombC = forms.ChoiceField(choices=TIPO_COMBUST_CAR,widget=forms.Select(attrs={'class':'form-control'}),required=False) #
     tCombM = forms.ChoiceField(choices=TIPO_COMBUST_MOTO,widget=forms.Select(attrs={'class':'form-control'}),required=False) #
